requirements/base.txt
django==1.9.4
mistune==0.7.1
pygments== 2.0.2
django-reversion==1.9.3
django-compressor==2.0
django-pagedown==0.1.0
